requirements.txt
absl-py==2.3.1
branca==0.8.2
certifi==2025.11.12
charset-normalizer==3.4.4
folium==0.20.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.0
pillow==12.0.0
requests==2.32.5
tqdm==4.67.1
urllib3==2.6.2
xyzservices==2025.11.0
